void autoImportScoresAndSettingsFromLocalShare()
This functions will check if the current application local dir is different of
"~/.local/share/libremines". This will occur when the user upgrade the application
from version 1.9.1 (or before) to the current 2.0.0 (or older).
In case the directories are different and that are user settings or high scores,
it will import those files by copying to the new data dir location.